[feature/save_view]: Add EmptyView
