+++ presentation/src/main/java/com/everymeal/presentation/ui/signup/UnivSelectViewModel.kt
@@ -1,8 +1,8 @@
 package com.everymeal.presentation.ui.signup
 
 import com.everymeal.presentation.base.BaseViewModel
-import com.everymeal.presentation.ui.signup.UnivSelectContract.UnivSelectEvent
 import com.everymeal.presentation.ui.signup.UnivSelectContract.UnivSelectEffect
+import com.everymeal.presentation.ui.signup.UnivSelectContract.UnivSelectEvent
 import com.everymeal.presentation.ui.signup.UnivSelectContract.UnivSelectState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
@@ -10,19 +10,22 @@ import javax.inject.Inject
 @HiltViewModel
 class UnivSelectViewModel @Inject constructor(
 
-): BaseViewModel<UnivSelectState, UnivSelectEffect, UnivSelectEvent>(
+) : BaseViewModel<UnivSelectState, UnivSelectEffect, UnivSelectEvent>(
     UnivSelectState()
 ) {
 
     override fun handleEvents(event: UnivSelectEvent) {
-        when(event) {
+        when (event) {
             is UnivSelectEvent.SelectButtonClicked -> {
                 sendEffect({ UnivSelectEffect.MoveToMain })
             }
+
             is UnivSelectEvent.SelectedUniv -> {
-                updateState { copy(
-                    selectedUniv = event.selectedUniv
-                ) }
+                updateState {
+                    copy(
+                        selectedUniv = event.selectedUniv
+                    )
+                }
             }
         }
     }
