+++ presentation/src/main/res/drawable/icon_save_empty.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="40dp"
+    android:height="40dp"
+    android:viewportWidth="40"
+    android:viewportHeight="40">
+  <path
+      android:pathData="M35,4.167H5C4.558,4.167 4.134,4.342 3.822,4.655C3.509,4.967 3.333,5.391 3.333,5.833V8.612H36.667V5.833C36.667,5.391 36.491,4.967 36.179,4.655C35.866,4.342 35.442,4.167 35,4.167ZM5.293,31.98C5.293,33.041 5.715,34.058 6.464,34.808C7.214,35.558 8.231,35.979 9.292,35.98H16.667V29.723C16.667,29.281 16.842,28.857 17.155,28.545C17.468,28.232 17.892,28.057 18.333,28.057H21.667C22.109,28.057 22.533,28.232 22.845,28.545C23.158,28.857 23.333,29.281 23.333,29.723V35.978H30.708C31.769,35.978 32.786,35.556 33.536,34.806C34.286,34.056 34.707,33.039 34.707,31.978V18.333H5.293V31.98ZM3.333,14.167C3.333,14.609 3.509,15.033 3.822,15.345C4.134,15.658 4.558,15.833 5,15.833H35C35.442,15.833 35.866,15.658 36.179,15.345C36.491,15.033 36.667,14.609 36.667,14.167V11.388H3.333V14.167Z"
+      android:fillColor="#B0B8C1"
+      android:fillType="evenOdd"/>
+</vector>
